requirements.txt
aiofiles==24.1.0
aiogram==3.18.0
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
annotated-types==0.7.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
async-timeout==5.0.1
asyncpg==0.30.0
attrs==25.1.0
certifi==2025.1.31
cffi==1.17.1
frozenlist==1.5.0
idna==3.10
magic-filter==1.0.12
minio==7.2.15
multidict==6.1.0
propcache==0.3.0
pycparser==2.22
pycryptodome==3.21.0
pydantic==2.10.6
pydantic_core==2.27.2
python-dotenv==1.0.1
typing_extensions==4.12.2
urllib3==2.3.0
yarl==1.18.3
